requirements.txt
cupy-cuda11x==13.6.0
fastrlock==0.8.3
numpy==2.2.6
pillow==12.0.0
